+++ apollo-router/src/router.rs
@@ -32,9 +32,9 @@ use self::Event::NoMoreSchema;
 use self::Event::Shutdown;
 use self::Event::UpdateConfiguration;
 use self::Event::UpdateSchema;
-use crate::axum_http_server_factory::make_axum_router;
-use crate::axum_http_server_factory::AxumHttpServerFactory;
-use crate::axum_http_server_factory::ListenAddrAndRouter;
+use crate::axum_factory::make_axum_router;
+use crate::axum_factory::AxumHttpServerFactory;
+use crate::axum_factory::ListenAddrAndRouter;
 use crate::configuration::Configuration;
 use crate::configuration::ListenAddr;
 use crate::plugin::DynPlugin;
